requirements.txt
streamlit
pymupdf
langchain
chromadb
tiktoken
openai
requests
python-dotenv
google-generativeai
google-cloud-aiplatform
google-genai 